Revert "clean code"
This reverts commit 54976fbf637b11ceec0dca37ad54758054b4d48f.

+++ src/main/java/cn/nukkit/network/protocol/SetScoreboardIdentityPacket.java
@@ -47,7 +47,7 @@ public class SetScoreboardIdentityPacket extends DataPacket {
     
     public static class SetScoreboardIdentityEntry {
 
-        public long scoreboardId;
+        public int scoreboardId;
         public long entityUniqueId;
     }
 }
